+++ src/main/resources/db/migration/V5.0__ticket_table_create.sql
@@ -1,7 +1,7 @@
 CREATE TABLE ticket
 (
     id                    BIGINT(20) AUTO_INCREMENT NOT NULL,
-    price                 DECIMAL(10, 2)            NOT NULL,
+    price                 FLOAT                     NOT NULL,
     date_of_departure     DATETIME                  NOT NULL,
     reservation_id        BIGINT,
     one_way_route_id      BIGINT,
